More 64-bit compatibility, mostly changing "long" to "int".
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1600 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/base/misc/dump.c
@@ -168,7 +168,7 @@ show_ints(int min, int max)
 #error MAX_SELECTORS needs to be 8192
 #endif
 
-#define GetSegmentBaseAddress(s)	(Segments[(s) >> 3].base_addr)
+#define GetSegmentBaseAddress(s)	((unsigned long)Segments[(s) >> 3].base_addr)
 #define IsSegment32(s)			(Segments[(s) >> 3].is_32)
 
 char *DPMI_show_state(struct sigcontext_struct *scp)
